requirements.txt
# 🌾 AI Farmer Credit Intelligence Platform
# Enterprise Production Dependencies

# Core ML & Data Processing
pandas>=1.5.3
numpy>=1.24.3
scikit-learn>=1.2.2
joblib>=1.2.0

# Web Framework & API
Flask>=2.3.2
Flask-CORS>=4.0.0
Jinja2>=3.1.2

# 🔐 Authentication & Security
Werkzeug>=2.3.4
itsdangerous>=2.1.2

# Database & Storage
sqlalchemy>=2.0.15
psycopg2-binary>=2.9.6

# Development & Utilities
python-dotenv>=1.0.0

# Optional: Advanced Analytics
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.14.1

# Optional: Streamlit for Demo Interface
streamlit>=1.23.1

# Security & Production 
gunicorn>=20.1.0
python-multipart>=0.0.6
